@@ -6,8 +6,7 @@ publisher:
 homepage: https://github.com/songyiYu/all_sensors
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.0 <2.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
